fixed merge conflicts

+++ sonar-polymorphic-views/src/main/java/be/kuleuven/cs/oss/datautils/Position.java
@@ -8,19 +8,19 @@ package be.kuleuven.cs.oss.datautils;
 
 public class Position {
 	
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
-	public Position(double x, double y){
+	public Position(int x, int y){
 		this.x = x;
 		this.y = y;
 	}
 	
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
